+++ src/pages/configuration/config-list/config-list.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import {connect} from 'react-redux';
 import {useHistory} from 'react-router-dom';
-import {reduxUtils, backend, templateUtils} from '@utilities';
+import {reduxUtils, templateUtils, localIdUtils, backend} from '@utilities';
 import {types} from '@types';
 import VisualConfigList from './visual-config-list';
+import RemoveSurveyPopup from './remove-survey-popup';
+import DuplicateSurveyPopup from './duplicate-survey-popup';
+
 interface Props {
     account: types.Account;
     accessToken: types.AccessToken;
@@ -12,6 +15,7 @@ interface Props {
     openMessage(messageId: types.MessageId): void;
     openModal(title: string, children: React.ReactNode): void;
     closeModal(): void;
+    removeConfig(surveyName: string): void;
     addConfig(config: types.SurveyConfig): void;
 }
 function ConfigList(props: Props) {
@@ -36,11 +40,81 @@ function ConfigList(props: Props) {
         );
     }
 
+    const removeSurvey = (config: types.SurveyConfig) => () => {
+        function success() {
+            props.removeConfig(config.survey_name);
+            props.closeModal();
+            history.push('/configurations');
+            props.openMessage('success-survey-removed');
+        }
+
+        function error() {
+            props.closeModal();
+            props.openMessage('error-server');
+        }
+
+        backend.deleteSurvey(
+            props.account,
+            props.accessToken,
+            config.survey_name,
+            success,
+            error,
+        );
+    };
+
+    const duplicateSurvey =
+        (config: types.SurveyConfig) => (newSurveyName: string) => {
+            const newConfig = localIdUtils.remove.survey({
+                ...config,
+                survey_name: newSurveyName,
+                draft: true,
+                fields: config.fields.map((f, i) => ({...f, identifier: i})),
+            });
+
+            const success = () => {
+                props.addConfig(newConfig);
+                props.closeModal();
+                history.push(`/configuration/${newSurveyName}`);
+                props.openMessage('success-survey-duplicated');
+            };
+            const error = (code: 400 | 401 | 422 | 500) => {};
+
+            backend.createSurvey(
+                props.account,
+                props.accessToken,
+                newConfig,
+                success,
+                error,
+            );
+        };
+
+    const openRemoveModal = (config: types.SurveyConfig) => () => {
+        props.openModal(
+            'Remove this survey permanently?',
+            <RemoveSurveyPopup
+                removeSurvey={removeSurvey(config)}
+                closeModal={props.closeModal}
+            />,
+        );
+    };
+
+    const openDuplicateModal = (config: types.SurveyConfig) => () => {
+        props.openModal(
+            'Duplicate this survey',
+            <DuplicateSurveyPopup
+                thisConfig={config}
+                duplicateSurvey={duplicateSurvey(config)}
+            />,
+        );
+    };
+
     return (
         <VisualConfigList
             configs={props.configs}
             addSurvey={addSurvey}
             account={props.account}
+            openRemoveModal={openRemoveModal}
+            openDuplicateModal={openDuplicateModal}
         />
     );
 }
@@ -54,6 +128,7 @@ const mapDispatchToProps = (dispatch: any) => ({
     openMessage: reduxUtils.dispatchers.openMessage(dispatch),
     openModal: reduxUtils.dispatchers.openModal(dispatch),
     closeModal: reduxUtils.dispatchers.closeModal(dispatch),
+    removeConfig: reduxUtils.dispatchers.removeConfig(dispatch),
     addConfig: reduxUtils.dispatchers.addConfig(dispatch),
 });
 export default connect(mapStateToProps, mapDispatchToProps)(ConfigList);
