+++ src/pages/configuration/config-list/duplicate-survey-popup.tsx
