+++ src/pages/configuration/config-list/remove-survey-popup.tsx
